project vocaldocs/CodeBuild_Artifacts/requirements.txt
pdf2image==1.17.0
Pillow
boto3